Refactor authentication system with modular components
- Create dedicated AuthProvider component for better separation of concerns
- Implement authService for centralized authentication logic
- Add apiUtils for standardized API request handling
- Update AuthContext to use the new service architecture
- Improve code organization with proper exports in index files
