+++ frontend/src/components/auth/index.ts
@@ -1,0 +1,5 @@
+// Exporter tous les composants d'authentification pour faciliter l'importation
+export { default as LoginForm } from './LoginForm';
+export { default as SignupForm } from './SignupForm';
+export { default as AuthProvider } from './AuthProvider';
+export { default as AuthLayout } from './AuthLayout';
